Update config.yml
players:
  -help.customcommandslist.txt

Self explained. Some mc mods have addressed this in their on documentation and or wiki.
For example portal menu  that uses the /help or a custom/help menu to show the player other commands for the server. This is most important when a server has a few dozen servers and a few connecting sub lobbies to get to the next server.